+++ app/src/main/java/madelyntav/c4q/nyc/googlecards/FlickrFragment.java
@@ -1,0 +1,293 @@
+package madelyntav.c4q.nyc.googlecards;
+
+
+import android.app.Fragment;
+import android.os.AsyncTask;
+import android.os.Bundle;
+import android.support.v4.widget.SwipeRefreshLayout;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ArrayAdapter;
+import android.widget.Button;
+import android.widget.EditText;
+import android.widget.GridView;
+import android.widget.ImageButton;
+import android.widget.LinearLayout;
+import android.widget.ListView;
+import android.widget.TextView;
+import android.widget.Toast;
+
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.MarkerOptions;
+
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.UnsupportedEncodingException;
+import java.net.HttpURLConnection;
+import java.net.URL;
+import java.net.URLEncoder;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class FlickrFragment extends Fragment {
+
+    private ImageButton mBtnFind, nameButton, largeNameButton, listButton, addCardButton;
+    private GoogleMap mMap;
+    private EditText etPlace;
+    private MapFragment mapFragment;
+    private ArrayList<String> list;
+    private EditText listEnter, enterAddress;
+    private EditText searchBar, nameText;
+    private Button homeButton, workButton;
+    private String homeAddress = "", workAddress = "", name = "";
+    private LinearLayout addressLayout, nameLayout, enterNameLayout;
+    private Button saveAddress;
+    private SwipeRefreshLayout swipeLayout;
+    private TextView nameView;
+    private ListView listView;
+    GridView mGridView;
+    ArrayAdapter<String> listAdapter;
+    ImageAdapter adapter;
+    View view;
+
+
+    public FlickrFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        view= inflater.inflate(R.layout.fragment_flickr, container, false);
+
+
+        return view;
+    }
+
+
+    /**
+     * A class, to download Places from Geocoding webservice
+     */
+    private class DownloadTask extends AsyncTask<String, Integer, String> {
+
+        String data = null;
+
+        // Invoked by execute() method of this object
+        @Override
+        protected String doInBackground(String... url) {
+            try {
+                data = downloadUrl(url[0]);
+            } catch (Exception e) {
+                Log.d("Background Task", e.toString());
+            }
+            return data;
+        }
+
+        // Executed after the complete execution of doInBackground() method
+        @Override
+        protected void onPostExecute(String result) {
+
+            // Instantiating ParserTask which parses the json data from Geocoding webservice
+            // in a non-ui thread
+            ParserTask parserTask = new ParserTask();
+
+            // Start parsing the places in JSON format
+            // Invokes the "doInBackground()" method of the class ParseTask
+            parserTask.execute(result);
+        }
+    }
+
+    /**
+     * A class to parse the Geocoding Places in non-ui thread
+     */
+    class ParserTask extends AsyncTask<String, Integer, List<HashMap<String, String>>> {
+
+        JSONObject jObject;
+
+        // Invoked by execute() method of this object
+        @Override
+        protected List<HashMap<String, String>> doInBackground(String... jsonData) {
+
+            List<HashMap<String, String>> places = null;
+            GeocodeJSONParser parser = new GeocodeJSONParser();
+
+            try {
+                jObject = new JSONObject(jsonData[0]);
+
+                /** Getting the parsed data as a an ArrayList */
+                places = parser.parse(jObject);
+
+            } catch (Exception e) {
+                Log.d("Exception", e.toString());
+            }
+            return places;
+        }
+
+        // Executed after the complete execution of doInBackground() method
+        @Override
+        protected void onPostExecute(List<HashMap<String, String>> list) {
+
+            // Clears all the existing markers
+            mMap.clear();
+
+            for (int i = 0; i < list.size(); i++) {
+
+                // Creating a marker
+                MarkerOptions markerOptions = new MarkerOptions();
+
+                // Getting a place from the places list
+                HashMap<String, String> hmPlace = list.get(i);
+
+                // Getting latitude of the place
+                double lat = Double.parseDouble(hmPlace.get("lat"));
+
+                // Getting longitude of the place
+                double lng = Double.parseDouble(hmPlace.get("lng"));
+
+                // Getting name
+                String name = hmPlace.get("formatted_address");
+
+                LatLng latLng = new LatLng(lat, lng);
+
+                // Setting the position for the marker
+                markerOptions.position(latLng);
+
+                // Setting the title for the marker
+                markerOptions.title(name);
+
+                // Placing a marker on the touched position
+                mMap.addMarker(markerOptions);
+
+                // Locate the first location
+                if (i == 0)
+                    mMap.animateCamera(CameraUpdateFactory.newLatLng(latLng));
+            }
+        }
+    }
+
+
+    public void findLocation(String location) {
+        if (location == null || location.equals("")) {
+            Toast.makeText(getActivity(), "No Place is entered", Toast.LENGTH_SHORT).show();
+            return;
+        }
+
+        String url = "https://maps.googleapis.com/maps/api/geocode/json?";
+
+        try {
+            // encoding special characters like space in the user input place
+            location = URLEncoder.encode(location, "utf-8");
+        } catch (UnsupportedEncodingException e) {
+            e.printStackTrace();
+        }
+
+        String address = "address=" + location;
+
+        String sensor = "sensor=false";
+
+        // url , from where the geocoding data is fetched
+        url = url + address + "&" + sensor;
+
+        // Instantiating DownloadTask to get places from Google Geocoding service
+        // in a non-ui thread
+        DownloadTask downloadTask = new DownloadTask();
+
+        // Start downloading the geocoding places
+        downloadTask.execute(url);
+    }
+
+
+    // FLICKR CARD CODE
+    private void initializeViews() {
+        mGridView = (GridView) view.findViewById(R.id.gridView);
+    }
+
+
+    private class AsyncLoading extends AsyncTask<Void, Void, List<String>> {
+
+        @Override
+        protected List<String> doInBackground(Void... params) {
+            // TODO : Step 3 - by using FlickrGetter.java, get latest 20 images' Urls from Flickr and return the result.
+
+
+            try {
+                return new FlickrGetter().getBitmapList();
+            } catch (JSONException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            return null;
+        }
+
+        @Override
+        protected void onPostExecute(List<String> imageList) {
+            // TODO : Step 5 - Now we have ImageAdapter and the data(list), post the picture!
+
+            adapter = new ImageAdapter(getActivity(), imageList);
+            mGridView.setAdapter(adapter);
+
+        }
+    }
+    private String downloadUrl(String strUrl) throws IOException {
+        String data = "";
+        InputStream iStream = null;
+        HttpURLConnection urlConnection = null;
+        try {
+            URL url = new URL(strUrl);
+            // Creating an http connection to communicate with url
+            urlConnection = (HttpURLConnection) url.openConnection();
+
+            // Connecting to url
+            urlConnection.connect();
+
+            // Reading data from url
+            iStream = urlConnection.getInputStream();
+
+            BufferedReader br = new BufferedReader(new InputStreamReader(iStream));
+
+            StringBuffer sb = new StringBuffer();
+
+            String line = "";
+            while ((line = br.readLine()) != null) {
+                sb.append(line);
+            }
+
+            data = sb.toString();
+            br.close();
+
+        } catch (Exception e) {
+            Log.d("Exception with url", e.toString());
+        } finally {
+            iStream.close();
+            urlConnection.disconnect();
+        }
+
+        return data;
+
+
+    }
+
+}
+
+
+
+
+
+
