+++ app/src/main/res/layout-land/activity_main.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="horizontal" android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <FrameLayout
+        android:layout_width="wrap_content"
+        android:layout_weight="1"
+        android:layout_height="fill_parent"
+        android:id="@+id/first_frag"/>
+
+    <FrameLayout
+        android:layout_width="wrap_content"
+        android:layout_weight="1"
+        android:layout_height="fill_parent"
+        android:id="@+id/second_frag"/>
+
+
+</LinearLayout>
