+++ app/src/main/res/layout/fragment_flickr.xml
@@ -1,0 +1,32 @@
+<android.support.v7.widget.CardView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:id="@+id/flickrCard"
+    tools:context="madelyntav.c4q.nyc.googlecards.FlickrFragment">
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+            <ImageView
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center_horizontal"
+                android:padding="5dp"
+                android:src="@drawable/flickr" />
+
+            <GridView
+                android:id="@+id/gridView"
+                android:layout_width="match_parent"
+                android:layout_height="300dp"
+                android:layout_gravity="center"
+                android:layout_margin="4dp"
+                android:numColumns="2"
+                android:padding="4dp" />
+
+        </LinearLayout>
+
+</android.support.v7.widget.CardView>
